chore: Extract runtime/task_locks.rs from orbit_tool_host [ORB-10396]
Planned-By: codex
Agent-Model: gpt-5.6-terra
