+++ crates/orbit-core/src/runtime/mod.rs
@@ -21,6 +21,7 @@ mod resolve;
 pub mod run_audit;
 pub(crate) mod run_input;
 mod task_block_on_run_failure;
+mod task_locks;
 mod task_records;
 mod task_reservation_cleanup;
 pub(crate) mod tool_exec;
