+++ crates/orbit-core/src/runtime/orbit_tool_host/tests/mod.rs
@@ -1,4 +1,3 @@
 mod learning_tools;
 mod state_tools;
-mod task_locks;
 mod task_tools;
