+++ crates/orbit-core/src/runtime/v2_host/pipeline_actions.rs
@@ -8,7 +8,7 @@ use serde_json::Value;
 
 use crate::OrbitRuntime;
 use crate::command::job::JobRunListParams;
-use crate::runtime::orbit_tool_host::parse_task_ids;
+use crate::runtime::task_locks::parse_task_ids;
 
 pub(super) fn validate_bundles(action: &str, input: &Value) -> Result<Value, DispatchError> {
     let bundles_raw = input
